+++ src/components/Label/labelChangelogData.ts
@@ -10,11 +10,20 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      'Changes the font weight for the `(required)` label to `"medium"`.',
+    ],
+  },
+  {
     date: "2024-09-19",
     version: "3.3.2",
     type: "Update",
     affects: ["Styles"],
-    notes: ["Changes 'r' in '(required)' label from upper- to lowercase"],
+    notes: ["Changes 'r' in '(required)' label from upper- to lowercase."],
   },
   {
     date: "2024-07-25",